requirements.txt
Flask
pytesseract
opencv-python-headless
requests
pillow
beautifulsoup4
numpy
gunicorn
